requirements.txt
beautifulsoup4
lxml
python-dateutil
requests
streamlit